pyproject.toml
[project]
name = "cc-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.3.3",
    "requests>=2.32.5",
    "google-auth>=2.37.0",
    "google-auth-oauthlib>=1.2.1",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.156.0",
    "mastodon-py>=1.8.1",
    "python-dotenv>=1.0.1",
]
